+++ app/src/androidTest/java/com/averycorp/prismtask/CloudIdOrphanHealerTier1Test.kt
@@ -92,6 +92,7 @@ class CloudIdOrphanHealerTier1Test {
             medicationSlotDao = database.medicationSlotDao(),
             medicationSlotOverrideDao = database.medicationSlotOverrideDao(),
             medicationTierStateDao = database.medicationTierStateDao(),
+            medicationMarkDao = database.medicationMarkDao(),
             logger = logger
         )
     }